pyproject.toml
[project]
name = "test-analysis-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "google-genai==0.5.0",
    "PyYAML==6.0.2",
    "python-dotenv==1.0.1",
    "uvicorn[standard]==0.30.6",
]
